Fix Total findings excluding alias
This commit fixes the calculation of total vulnerabilities displayed in the badge
by excluding duplicates that occur within the same component.
Previously, the badge did not accurately reflect the total number of unique vulnerabilities.

To address this issue, we added logic to count each vulnerability only once, and avoid
vulnerabilities already mentioned in aliases. By doing this, we are able to avoid counting the same vulnerability
multiple times within the same component.

Signed-off-by: Andres Tito <andres.tito@rohde-schwarz.com>